requirements.txt
requests
pandas
numpy
beautifulsoup4
wordcloud
matplotlib
ipadic
mecab-python3
